requirements.txt
alpaca-py==0.40.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
Authlib==1.5.2
autopep8==2.3.2
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
executing==2.2.0
Flask==3.1.0
fonttools==4.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibapi @ file:///home/moshe/Downloads/twsapi_macunix.1036.01/IBJts/source/pythonclient/dist/ibapi-10.36.1-py3-none-any.whl#sha256=9cd85ed4a063b8c4b9b7dbe5ab16c167ad08dfe683320eb729419c3017c08dc1
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mplcursors==0.6
msgpack==1.1.0
multiprocess==0.70.18
mypy==1.15.0
mypy_extensions==1.1.0
numba==0.61.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pandas-stubs==2.2.3.250308
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
schwab-py==1.5.0
scipy==1.15.2
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
sseclient-py==1.8.0
stack-data==0.6.3
stumpy==1.13.0
traitlets==5.14.3
types-pytz==2025.2.0.20250326
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
